backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3
bcrypt==4.1.3
python-jose==3.3.0
python-multipart==0.0.12
google-auth==2.35.0
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
requests==2.31.0